+++ frontend/src/pages/RestaurantDetailPage/ImageViewer.tsx
@@ -1,0 +1,18 @@
+import { Restaurant } from '~/@types/restaurant.types';
+import ImageCarousel from '~/components/@common/ImageCarousel';
+import ImageGrid from '~/components/@common/ImageGrid';
+import useMediaQuery from '~/hooks/useMediaQuery';
+
+interface ImageViewerProps {
+  images: Restaurant['images'];
+}
+
+function ImageViewer({ images }: ImageViewerProps) {
+  const { isMobile } = useMediaQuery();
+
+  if (isMobile) return <ImageCarousel type="list" images={images} />;
+
+  return <ImageGrid images={images.map(({ name: url, author, sns }) => ({ waterMark: author, url, sns }))} />;
+}
+
+export default ImageViewer;
